feat(site): add language-switcher + update testimonials + footer
